// block-performance-1.0.0/src/library/file-system/directory-ls/type.ts
export interface Input {
    dir: string;
    recursive: boolean;
}

export interface OutputEmpty {
    directory_list: { name: string; fullname: string }[];
}

export interface OutputOk {
    directory_list: { name: string; fullname: string }[];
}

export interface OutputFail {}

export interface Callback<R> {
    empty: (output: OutputEmpty) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;